# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Islamic Andolon Project API",
        "description": "API documentation",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local server"
        }
    ],
    "paths": {
...
        "/api/profiles": {
            "get": {
                "tags": [
                    "Profiles"
                ],
                "summary": "List all profiles",
                "operationId": "43d12304c3e621ecce8cc9ccd4b3738e",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "post": {
                "tags": [
                    "Profiles"
                ],
                "summary": "Create a new profile",
                "operationId": "00abb5f24a20112b161d17e9a5c38a37",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id",
                                    "name",
                                    "mobile",
                                    "seat_id",
                                    "division_id",
                                    "district_id",
                                    "upazila_id",
                                    "union_id"
                                ],
                                "properties": {
                                    "user_id": {
                                        "type": "integer"
                                    },
                                    "name": {
                                        "type": "string"
                                    },
                                    "mobile": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "seat_id": {
                                        "type": "integer"
                                    },
                                    "division_id": {
                                        "type": "integer"
                                    },
                                    "district_id": {
                                        "type": "integer"
                                    },
                                    "upazila_id": {
                                        "type": "integer"
                                    },
                                    "union_id": {
                                        "type": "integer"
                                    },
                                    "message": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Profile created successfully"
                    }
                }
            }
        },
        "/api/profiles/{id}": {
            "get": {
                "tags": [
                    "Profiles"
                ],
                "summary": "Get profile details",
                "operationId": "bdbf886f973e16c2be2a83e76a804833",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Profile ID",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Profile details"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Profiles"
                ],
                "summary": "Update profile",
                "operationId": "c878463709815b7780e0e3b989a06bec",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Profile ID",
                        "required": true
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "user_id": {
                                        "type": "integer"
                                    },
                                    "name": {
                                        "type": "string"
                                    },
                                    "mobile": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "seat_id": {
                                        "type": "integer"
                                    },
                                    "division_id": {
                                        "type": "integer"
                                    },
                                    "district_id": {
                                        "type": "integer"
                                    },
                                    "upazila_id": {
                                        "type": "integer"
                                    },
                                    "union_id": {
                                        "type": "integer"
                                    },
                                    "message": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Profile updated successfully"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Profiles"
                ],
                "summary": "Delete profile",
                "operationId": "f1f615f4e7966b63084409a8d09eed5d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Profile ID",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Deleted successfully"
                    }
                }
            }
        },
        "/api/seats": {
            "get": {
                "tags": [
                    "Seats"
                ],
                "summary": "List all seats",
                "operationId": "09a0c40bbde096321060e7db24676668",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "post": {
                "tags": [
                    "Seats"
                ],
                "summary": "Create a new seat",
                "operationId": "2c95763f806f23800eeca9b867af7147",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "seat_no",
                                    "seat_name",
                                    "candidate_name",
                                    "area"
                                ],
                                "properties": {
                                    "seat_no": {
                                        "type": "string"
                                    },
                                    "seat_name": {
                                        "type": "string"
                                    },
                                    "candidate_name": {
                                        "type": "string"
                                    },
                                    "area": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Seat created successfully"
                    }
                }
            }
        },
        "/api/seats/{id}": {
            "get": {
                "tags": [
                    "Seats"
                ],
                "summary": "Get seat details",
                "operationId": "ce05d024b00285d14e8ba4c613b1308c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Seat ID",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Seat details"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Seats"
                ],
                "summary": "Update seat",
                "operationId": "b6acc6adda5646bb299c4a0f63a6be49",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Seat ID",
                        "required": true
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "seat_no": {
                                        "type": "string"
                                    },
                                    "seat_name": {
                                        "type": "string"
                                    },
                                    "candidate_name": {
                                        "type": "string"
                                    },
                                    "area": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Seat updated successfully"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Seats"
                ],
                "summary": "Delete seat",
                "operationId": "d030c02dbd229750cf4ea0a6a0cf3f33",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Seat ID",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Deleted successfully"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}